Decouple instrumenting from logging
